+++ example/lib/main.dart
@@ -35,12 +35,14 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: Text('OpenGL plugin example'),
+          title: Text('Flutter OpenGL test'),
         ),
-        body: new Center(
+        body: Center(
           child: Column(
             children: <Widget>[
               SizedBox(height: 8.0,),
+              Text("https://github.com/alnitak/flutter_opengl"),
+              SizedBox(height: 6.0,),
               Container(
                 width: _width,
                 height: _height,
@@ -48,82 +50,92 @@ class _MyAppState extends State<MyApp> {
                     ? Texture(textureId: controller.textureId)
                     : Container(),
               ),
-              Padding(
-                padding: const EdgeInsets.only(top: 16.0),
-                child: Text(notes),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  RaisedButton(
-                    child: Text("stop"),
-                    color: Colors.red,
-                    textColor: Colors.white,
-                    onPressed: () {
-                      stopNDKController();
-                    },
-                  ),
-                  SizedBox(width: 20),
-                  RaisedButton(
-                    child: Text("re-create"),
-                    color: Colors.redAccent,
-                    textColor: Colors.white,
-                    onPressed: () {
-                      initializeNDKController();
-                    },
-                  ),
-                ],
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  RaisedButton(
-                    child: Text("OpenGL ES 1\nfunc 1"),
-                    onPressed: () {
-                      setNDKRenderFuncController(0);
-                      notes = "";
-                      setState(() {});
-                    },
-                  ),
-                  SizedBox(width: 6),
-                  RaisedButton(
-                    child: Text("OpenGL ES 2\nshader 1"),
-                    onPressed: () {
-                      setNDKRenderFuncController(1);
-                      notes = "COMPUTATIONAL INTENSIVE\nhttps://www.shadertoy.com/view/3l23Rh";
-                      setState(() {});
-                    },
-                  ),
-                ],
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  RaisedButton(
-                    child: Text("OpenGL ES 2\nshader 2"),
-                    onPressed: () {
-                      setNDKRenderFuncController(2);
-                      notes = "https://www.shadertoy.com/view/XsXXDn";
-                      setState(() {});
-                    },
-                  ),
-                  SizedBox(width: 6),
-                  RaisedButton(
-                    child: Text("OpenGL ES 2\nshader 3"),
-                    onPressed: () {
-                      setNDKRenderFuncController(3);
-                      notes = "https://www.shadertoy.com/view/ttlGDf";
-                      setState(() {});
-                    },
-                  ),
-                ],
-              ),
+              _buildControls(),
             ],
           ),
         ),
       ),
     );
   }
+
+  Widget _buildControls() {
+    return
+        Column(
+          children: <Widget>[
+            Padding(
+              padding: const EdgeInsets.only(top: 16.0),
+              child: Text(notes),
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                RaisedButton(
+                  child: Text("stop"),
+                  color: Colors.red,
+                  textColor: Colors.white,
+                  onPressed: () {
+                    stopNDKController();
+                  },
+                ),
+                SizedBox(width: 20),
+                RaisedButton(
+                  child: Text("re-create"),
+                  color: Colors.green,
+                  textColor: Colors.white,
+                  onPressed: () {
+                    initializeNDKController();
+                  },
+                ),
+              ],
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                RaisedButton(
+                  child: Text("OpenGL ES 1\nfunc 1"),
+                  onPressed: () {
+                    setNDKRenderFuncController(0);
+                    notes = "";
+                    setState(() {});
+                  },
+                ),
+                SizedBox(width: 6),
+                RaisedButton(
+                  child: Text("OpenGL ES 2\nshader 1"),
+                  onPressed: () {
+                    setNDKRenderFuncController(1);
+                    notes = "COMPUTATIONAL INTENSIVE\nhttps://www.shadertoy.com/view/3l23Rh";
+                    setState(() {});
+                  },
+                ),
+              ],
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                RaisedButton(
+                  child: Text("OpenGL ES 2\ntexture shader 2"),
+                  onPressed: () {
+                    setNDKRenderFuncController(2);
+                    notes = "https://www.shadertoy.com/view/llj3Dz";
+                    setState(() {});
+                  },
+                ),
+                SizedBox(width: 6),
+                RaisedButton(
+                  child: Text("OpenGL ES 2\nshader 3"),
+                  onPressed: () {
+                    setNDKRenderFuncController(3);
+                    notes = "https://www.shadertoy.com/view/ttlGDf";
+                    setState(() {});
+                  },
+                ),
+              ],
+            ),
+          ],
+        );
+  }
+
 
   initializeNDKController() async {
     await controller.initializeNDK(0, _width, _height);