+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.0.2
 
-* TODO: Describe initial release.
+* code rearranged.
+* every shader can now be written separately subclassing the Shader class and passing it to the renderer.
+* added function to load texture image from assets folder with Android. Since this uses java, it should be rewritten for iOS.
+* added a texture shader example.